+++ DofusProtocol/Types/Types/game/achievement/AchievementRewardable.cs
@@ -1,6 +1,6 @@
 
 
-// Generated on 10/28/2014 16:37:59
+// Generated on 12/29/2014 21:14:13
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -18,13 +18,13 @@ namespace Stump.DofusProtocol.Types
         }
         
         public short id;
-        public short finishedlevel;
+        public byte finishedlevel;
         
         public AchievementRewardable()
         {
         }
         
-        public AchievementRewardable(short id, short finishedlevel)
+        public AchievementRewardable(short id, byte finishedlevel)
         {
             this.id = id;
             this.finishedlevel = finishedlevel;
@@ -33,7 +33,7 @@ namespace Stump.DofusProtocol.Types
         public virtual void Serialize(IDataWriter writer)
         {
             writer.WriteShort(id);
-            writer.WriteShort(finishedlevel);
+            writer.WriteByte(finishedlevel);
         }
         
         public virtual void Deserialize(IDataReader reader)
@@ -41,14 +41,14 @@ namespace Stump.DofusProtocol.Types
             id = reader.ReadShort();
             if (id < 0)
                 throw new Exception("Forbidden value on id = " + id + ", it doesn't respect the following condition : id < 0");
-            finishedlevel = reader.ReadShort();
+            finishedlevel = reader.ReadByte();
             if (finishedlevel < 0 || finishedlevel > 200)
                 throw new Exception("Forbidden value on finishedlevel = " + finishedlevel + ", it doesn't respect the following condition : finishedlevel < 0 || finishedlevel > 200");
         }
         
         public virtual int GetSerializationSize()
         {
-            return sizeof(short) + sizeof(short);
+            return sizeof(short) + sizeof(byte);
         }
         
     }
