+++ DofusProtocol/Types/Types/game/context/fight/GameFightFighterMonsterLightInformations.cs
@@ -1,6 +1,6 @@
 
 
-// Generated on 10/28/2014 16:38:02
+// Generated on 12/29/2014 21:14:24
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -23,7 +23,7 @@ namespace Stump.DofusProtocol.Types
         {
         }
         
-        public GameFightFighterMonsterLightInformations(bool sex, bool alive, int id, int wave, short level, sbyte breed, short creatureGenericId)
+        public GameFightFighterMonsterLightInformations(bool sex, bool alive, int id, sbyte wave, short level, sbyte breed, short creatureGenericId)
          : base(sex, alive, id, wave, level, breed)
         {
             this.creatureGenericId = creatureGenericId;
