+++ DofusProtocol/Types/Types/game/shortcut/ShortcutObjectItem.cs
@@ -1,6 +1,6 @@
 
 
-// Generated on 10/28/2014 16:38:06
+// Generated on 12/29/2014 21:14:45
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -24,7 +24,7 @@ namespace Stump.DofusProtocol.Types
         {
         }
         
-        public ShortcutObjectItem(int slot, int itemUID, int itemGID)
+        public ShortcutObjectItem(sbyte slot, int itemUID, int itemGID)
          : base(slot)
         {
             this.itemUID = itemUID;
